load titles from metadata in examples page
